pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily_digest_bot"
version = "0.1.0"
description = "Daily news digest CLI that fetches RSS, ranks stories, renders HTML, and emails via Gmail SMTP."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "feedparser>=6.0",
  "PyYAML>=6.0",
  "requests>=2.31",
  "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
